Moving routes functions out of server.ts and wrapping up render functionality in same place.
